s/helpers/utilities, add ride reducer test
