+++ app/javascript/driving/utilities/fetch.js
